Read a Dataframe from Clipboard

Data types are double, integer, numeric, and character. All factors are converted to character before the dataframe is returned. If header is false, the dataframe is returned with all the rows and column names in the pattern "Vindex," where "index" is the position of the column.

read_clipboard(header = TRUE, log = "DEPRECATED", log_details = "")

Arguments

header
True if data has a header.

log
(Deprecated) Log this function call to a log.md file?

log_details
(Deprecated) If log is TRUE, additional notes to include with the log entry.

See also

Examples

library(tidyverse)

# Test Data
test_data <-
        tibble(A = 1:3,
               B = c("Apples", "Bananas", "Oranges"),
               C = factor(c("Apples", "Bananas", "Oranges")),
               D = runif(3))

# Copy to Clipboard
copy_cb(data = test_data)

# Read Clipboard
read_cb()

#>   A       B       C          D
#> 1 1  Apples  Apples 0.03424133
#> 2 2 Bananas Bananas 0.32038573
#> 3 3 Oranges Oranges 0.40232824

# Check Datatypes
sapply(test_data, FUN = class)

#>           A           B           C           D 
#>   "integer" "character"    "factor"   "numeric" 

sapply(read_cb(), FUN = class) # `read_cb()` function mutates all factors to character before returning

#>           A           B           C           D 
#>   "integer" "character" "character"   "numeric" 

# Read Clipboard Shortcut for Headerless Data
read_cb_hl()

#>   V1      V2      V3                 V4
#> 1  A       B       C                  D
#> 2  1  Apples  Apples 0.0342413326725364
#> 3  2 Bananas Bananas  0.320385730825365
#> 4  3 Oranges Oranges  0.402328238356858

# Read Clipboard as a Tribble for Copy and Paste
read_cb_trbl()

#> tibble::tribble(,
#> ~A,~B,~C,~D,
#> 1, "Apples", "Apples", 0.0342413326725364,
#> 2, "Bananas", "Bananas", 0.320385730825365,
#> 3, "Oranges", "Oranges", 0.402328238356858,
#> )

read_cb_trbl(header = FALSE)

#> tibble::tribble(,
#> ~V1,~V2,~V3,~V4,
#> "A", "B", "C", "D",
#> "1", "Apples", "Apples", "0.0342413326725364",
#> "2", "Bananas", "Bananas", "0.320385730825365",
#> "3", "Oranges", "Oranges", "0.402328238356858",
#> )

read_cb_trbl(quote = "'")

#> tibble::tribble(,
#> ~A,~B,~C,~D,
#> 1, 'Apples', 'Apples', 0.0342413326725364,
#> 2, 'Bananas', 'Bananas', 0.320385730825365,
#> 3, 'Oranges', 'Oranges', 0.402328238356858,
#> )

read_cb_trbl(header = FALSE, quote = "'")

#> tibble::tribble(,
#> ~V1,~V2,~V3,~V4,
#> 'A', 'B', 'C', 'D',
#> '1', 'Apples', 'Apples', '0.0342413326725364',
#> '2', 'Bananas', 'Bananas', '0.320385730825365',
#> '3', 'Oranges', 'Oranges', '0.402328238356858',
#> )